requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==4.0.1